First Commit
I have made the navigation so that i can get from pageto pagewith ease i aslo have put the g1 tag on the main topic in the section element and put the paragraphs in a p tag and have done the unordered list. As well as bolding the money cost so far everything seems to be working fine.