One-Third Commit
I started this section using the header element then i had to add a header since i needed to add a photo and a title of sorts. From there i go down and have to put all the paragraphs in p tags and the sub headers in a h2 tag so they stand out more and then I bolded the Diane's Run and started on the footer which was a copy and paste situation and adding the special characters with the id for it. On the info page i aslo linked the base and style sheet.