Final Commit
I worked on the faq page the index and the info pages. Finished all of them. Starting with the info page from there started most of the work. The instructions page was not helpful in the least with the explanations which is fine since we have a debug squad that are full of kind helpful people. So when i started on this i put the title linked the image put the alt text and started the h1 and added the id for top of page where a later element will use to bring us back tot he top if clicked on and filled in the h1 text. After that i made an ordered list i added li tags and put an a href tag inside and gave each subtopic an id so that when i click them it will take me to each section and that makes the traversing the faq much easier. Then at the bottom of all that i added a back to top text inside a a href nested inside a p tag that had the id of back to top.